+++ pyctr/type/config/save.py
@@ -26,6 +26,10 @@ class InvalidConfigSaveError(ConfigSaveError):
     """Config Save is corrupted."""
 
 
+class OutOfSpaceConfigSaveError(ConfigSaveError):
+    """Config Save generation ran out of space for data."""
+
+
 class BlockFlagsNotAllowed(ConfigSaveError):
     """Flags not allowed. Must be 8, 12, 10, or 14 (0x8, 0xC, 0xA, or 0xE)."""
 
@@ -59,16 +63,9 @@ class ConfigSaveReader:
     https://www.3dbrew.org/wiki/Config_Savegame
     """
 
-    __slots__ = ('blocks', 'data_offset')
-
-    data_offset: int
-    """
-    Offset of the block data region. This is only used when rebuilding the raw save.
-    If a config save is loaded, this is overwritten with the offset from that file.
-    """
+    __slots__ = ('blocks',)
 
     def __init__(self):
-        self.data_offset = 0x41E4
         self.blocks: Dict[int, BlockInfo] = {}
 
     def __bytes__(self):
@@ -78,17 +75,20 @@ class ConfigSaveReader:
         """
         Converts the object to a raw config save file.
 
-        The result may not be 1:1 identical to the original config file due to garbage data in unused parts, as well as
-        offsets shifting around due to Python not always preserving order. This shouldn't matter in practice though.
+        CFG adds new block from end to start of file for any block > 4 bytes.
+        Any block <= 4 bytes only get a block entry.
 
         :return: Raw config save data.
         """
         raw_entries = []
         raw_block_datas = []
 
-        # offset in the block entry is from the start of the file,
-        # so the first one has the offset equal the data region offset
-        current_offset = self.data_offset
+        # offset starts at the end of the file
+        # this decrements per each data block that is bigger than 4 bytes
+        current_offset = CONFIG_SAVE_SIZE
+
+        # header + block entries
+        data_offset_limit = 4 + len(self.blocks) * BLOCK_ENTRY_SIZE
 
         for block_id, block_entry in self.blocks.items():
             data_size = len(block_entry.data)
@@ -98,21 +98,42 @@ class ConfigSaveReader:
                 data_size.to_bytes(2, 'little'),
                 block_entry.flags.to_bytes(2, 'little')
             ]
+
             if data_size > 4:
+                current_offset -= data_size
+                if current_offset < data_offset_limit:
+                    raise OutOfSpaceConfigSaveError("Too much block data for this save!")
+
                 raw_entry_list[1] = current_offset.to_bytes(4, 'little')
-                current_offset += data_size
-                raw_block_datas.append(block_entry.data)
+                raw_block_datas.insert(0, block_entry.data)
             else:
                 raw_entry_list[1] = block_entry.data.ljust(4, b'\0')
-                print(raw_entry_list[1].hex(), data_size)
+
             raw_entries.append(b''.join(raw_entry_list))
 
-        return b''.join((
+        hdr_and_entries = b''.join((
             len(self.blocks).to_bytes(2, 'little'),
-            self.data_offset.to_bytes(2, 'little'),
-            b''.join(raw_entries).ljust(self.data_offset - 4, b'\0'),
-            *raw_block_datas
-        )).ljust(CONFIG_SAVE_SIZE, b'\0')
+            current_offset.to_bytes(2, 'little'),
+            *raw_entries
+        ))
+        blks_data = b''.join(raw_block_datas)
+
+        if len(hdr_and_entries) != data_offset_limit:
+            raise ConfigSaveError("Something failed while making bytes, unexpected length of header and entries. Likely coding bug.")
+
+        if current_offset != CONFIG_SAVE_SIZE - len(blks_data):
+            raise ConfigSaveError("Something failed while making bytes, invalid data offset. Likely coding bug.")
+
+        config = b''.join((
+            hdr_and_entries,
+            bytes(CONFIG_SAVE_SIZE - len(blks_data) - len(hdr_and_entries)),
+            blks_data
+        ))
+
+        if len(config) != CONFIG_SAVE_SIZE:
+            raise ConfigSaveError("Something failed while making bytes, invalid bytes len. Likely coding bug.")
+
+        return config
 
     def save(self, fn: 'FilePath'):
         """
@@ -133,14 +154,14 @@ class ConfigSaveReader:
         :param data: Block data.
         """
 
-        if flags not in ALLOWED_FLAGS:
-            raise BlockFlagsNotAllowed(flags)
-
-        if not flags:
+        if flags is None:
             if block_id in self.blocks:
                 flags = self.blocks[block_id].flags
             else:
                 flags = 0xE
+
+        if flags not in ALLOWED_FLAGS:
+            raise BlockFlagsNotAllowed(flags)
 
         self.blocks[block_id] = BlockInfo(flags=flags, data=data)
 
@@ -181,7 +202,13 @@ class ConfigSaveReader:
         entry_count = int.from_bytes(header_raw[0:2], 'little')
         data_offset = int.from_bytes(header_raw[2:4], 'little')
 
-        block_entries = raw_save[4:data_offset]
+        block_entries_roof = 4 + BLOCK_ENTRY_SIZE * entry_count
+        if block_entries_roof > data_offset:
+            raise InvalidConfigSaveError(f'Data offset overlapped with entry headers')
+
+        block_entries = raw_save[4:block_entries_roof]
+
+        last_offset = CONFIG_SAVE_SIZE
 
         def load_raw_block_entry(b: bytes):
             block_id = int.from_bytes(b[0:4], 'little')
@@ -192,15 +219,30 @@ class ConfigSaveReader:
                 block_data_offset = int.from_bytes(b[4:8], 'little')
                 block_data = raw_save[block_data_offset:block_data_offset + block_size]
             else:
+                block_data_offset = -1
                 block_data = b[4:4 + block_size]
 
-            return {'id': block_id, 'flags': block_flags, 'data': block_data}
+            return {
+                'id': block_id,
+                'flags': block_flags,
+                'data': block_data,
+                'offset': block_data_offset,
+                'size': block_size
+            }
+
+        def sanity_check_blk(last_off: int, block: dict) -> int:
+            if block['size'] <= 4:
+                return last_off
+            last_off -= block['size']
+            if last_off != block['offset'] or last_off < data_offset:
+                raise InvalidConfigSaveError(f'Save not sane! May be corrupted.')
+            return last_off
 
         cfg_save = cls()
-        cfg_save.data_offset = data_offset
         for x in range(entry_count):
             entry = block_entries[x * BLOCK_ENTRY_SIZE:(x + 1) * BLOCK_ENTRY_SIZE]
             block = load_raw_block_entry(entry)
+            last_offset = sanity_check_blk(last_offset, block)
             cfg_save.set_block(block['id'], block['data'], block['flags'])
 
         return cfg_save